Calculates a bootstrapped confidence interval and other stats for a sample.

A bootstrapped confidence interval for the desired estimator for the provided sample is calculated for a confidence level level. Other stats and parameters of the distribution and sample are also returned.

Usage

calculate_boot_stats(
  sample,
  rep,
  n = "auto",
  level = 0.95,
  estimator = "mean",
  seed = NULL,
  pass_dist = FALSE
)

Arguments

sample

A numeric vector to bootstrap

rep

A integer vector for number of replicates

n

A integer or character vector for the size of bootstrap samples

level

A numeric vector for the confidence level

estimator

A character vector containing one of the ("mean", "median", "var", "sd") estimators

seed

A integer vector as random seed. Can be NULL

pass_dist

A boolean vector to decide whether to return the bootstrapped distribution

Value

A named list

Examples

calculate_boot_stats(c(1, 2, 3, 4), 1000, level = 0.95, seed = 1)
#> $lower
#> 2.5% 
#> 1.25 
#> 
#> $upper
#> 97.5% 
#>   3.5 
#> 
#> $sample_mean
#> [1] 2.5
#> 
#> $std_err
#> [1] 0.5655478
#> 
#> $level
#> [1] 0.95
#> 
#> $sample_size
#> [1] 4
#> 
#> $n
#> [1] "auto"
#> 
#> $rep
#> [1] 1000
#> 
#> $estimator
#> [1] "mean"
#> 
calculate_boot_stats(c(1, 2, 3, 4), 1000, level = 0.95, estimator = "median",
seed = 1, pass_dist = TRUE)
#> $lower
#> 2.5% 
#>    1 
#> 
#> $upper
#> 97.5% 
#>     4 
#> 
#> $sample_median
#> [1] 2.5
#> 
#> $std_err
#> [1] 0.7612427
#> 
#> $level
#> [1] 0.95
#> 
#> $sample_size
#> [1] 4
#> 
#> $n
#> [1] "auto"
#> 
#> $rep
#> [1] 1000
#> 
#> $estimator
#> [1] "median"
#> 
#> $dist
#>    [1] 2.0 2.5 2.5 1.0 2.0 1.0 1.0 2.0 2.0 2.5 2.5 4.0 1.5 3.5 2.5 2.5 2.0 2.0
#>   [19] 3.0 3.0 3.0 2.5 2.0 2.5 1.5 4.0 1.5 2.0 2.0 3.5 3.0 2.5 2.5 2.0 3.0 3.5
#>   [37] 3.0 2.5 1.5 3.0 1.0 2.5 3.0 2.5 2.0 3.0 3.0 3.5 3.5 2.0 2.0 2.0 3.0 2.5
#>   [55] 3.5 3.0 2.5 2.0 2.0 2.0 3.5 3.5 1.5 2.0 3.5 1.0 1.0 2.0 1.0 2.0 2.0 3.0
#>   [73] 1.5 4.0 3.0 1.0 2.0 2.5 1.0 2.5 3.5 3.5 2.5 4.0 2.5 3.0 1.5 1.5 2.5 2.5
#>   [91] 2.5 3.0 3.0 3.0 1.5 3.0 2.0 3.0 2.5 1.5 1.5 3.5 3.0 3.0 1.5 2.5 1.5 2.0
#>  [109] 2.5 3.0 3.0 1.5 1.5 3.5 3.0 4.0 2.5 1.5 2.5 2.0 3.0 2.5 2.5 3.0 3.5 3.0
#>  [127] 2.0 1.0 3.0 3.0 3.0 2.0 2.5 1.0 1.0 2.0 2.5 2.5 3.0 3.5 1.5 1.0 2.5 2.5
#>  [145] 1.5 2.0 3.0 3.5 2.5 3.5 1.5 2.5 2.0 3.0 2.0 3.0 3.0 2.0 3.5 2.0 1.5 2.5
#>  [163] 3.5 3.5 4.0 2.5 2.5 2.0 4.0 1.5 2.5 2.0 3.0 2.5 4.0 4.0 1.0 1.5 3.5 3.5
#>  [181] 3.0 2.5 1.5 1.5 3.0 2.5 2.5 2.0 3.5 2.5 2.5 2.5 3.0 1.0 1.5 1.5 1.5 3.0
#>  [199] 1.0 2.0 3.5 3.5 2.0 1.0 4.0 3.0 2.5 2.5 2.0 3.5 3.5 2.0 2.5 1.5 2.0 2.0
#>  [217] 2.0 3.0 3.0 2.5 3.0 2.5 3.0 2.5 2.5 4.0 2.0 3.5 3.0 2.5 1.5 1.5 2.5 2.0
#>  [235] 2.5 2.5 1.5 2.0 2.0 2.5 2.5 1.5 4.0 1.5 3.0 2.5 3.0 1.0 2.0 3.0 3.5 3.5
#>  [253] 1.5 2.0 2.5 2.5 2.5 2.5 1.0 1.5 3.0 1.5 1.0 3.5 2.5 4.0 2.5 1.0 2.5 2.0
#>  [271] 3.0 2.5 1.0 1.5 3.5 2.0 2.5 2.5 3.5 2.5 3.0 2.5 2.5 2.5 2.5 1.5 3.0 2.5
#>  [289] 3.0 2.0 4.0 3.5 2.0 2.5 4.0 2.5 2.5 2.5 3.0 3.5 3.5 2.5 2.0 3.0 3.5 2.5
#>  [307] 1.0 2.0 3.5 2.0 2.5 3.0 1.5 1.5 4.0 3.5 2.0 1.5 2.5 1.5 3.0 2.0 2.0 2.5
#>  [325] 3.0 3.5 3.0 1.5 3.5 3.0 2.0 4.0 1.5 4.0 2.5 3.0 1.5 2.0 3.5 1.5 2.5 1.5
#>  [343] 3.5 3.0 2.5 3.0 2.0 3.0 3.0 2.0 2.5 2.5 2.0 3.5 2.5 2.0 2.0 1.5 3.0 2.5
#>  [361] 2.5 2.0 2.5 3.5 4.0 1.0 2.5 1.0 3.0 1.5 4.0 2.0 1.0 2.0 3.0 2.5 2.5 3.0
#>  [379] 2.0 3.5 2.0 3.0 3.5 3.0 2.5 1.5 3.0 3.5 2.0 3.0 3.0 3.0 2.0 1.0 2.5 3.0
#>  [397] 2.5 3.5 2.0 1.0 3.0 2.5 2.0 1.5 1.0 3.5 3.0 2.5 3.0 3.5 1.5 3.0 2.5 2.0
#>  [415] 3.0 3.0 3.0 2.5 2.0 1.5 2.5 3.5 3.0 2.5 2.5 3.0 3.0 1.5 2.5 1.5 2.0 1.0
#>  [433] 2.5 3.0 2.0 2.0 1.0 2.5 3.5 3.5 2.0 1.5 2.5 2.5 3.5 2.5 3.0 3.5 2.0 1.0
#>  [451] 2.5 2.5 1.5 4.0 3.5 2.5 1.5 2.5 2.5 2.0 4.0 3.0 3.5 3.5 4.0 2.5 2.0 2.5
#>  [469] 1.0 3.0 2.5 3.0 3.5 2.5 2.5 3.0 3.5 3.5 2.0 3.5 3.0 2.0 2.0 2.5 3.0 2.0
#>  [487] 2.0 2.5 2.5 2.0 2.0 2.5 2.5 3.0 2.0 2.5 3.5 2.0 3.0 2.5 3.0 2.5 3.5 2.5
#>  [505] 2.5 3.0 3.0 3.5 1.5 2.0 2.0 2.0 3.5 3.5 2.5 2.0 4.0 3.5 1.5 2.5 2.0 3.0
#>  [523] 2.0 2.0 1.5 2.0 1.5 1.5 3.0 3.5 2.0 1.5 4.0 1.5 2.5 2.0 2.5 2.5 2.5 1.5
#>  [541] 2.0 2.0 4.0 3.0 3.0 1.5 1.5 1.5 4.0 2.0 3.5 3.5 1.5 2.5 4.0 1.5 1.5 1.5
#>  [559] 2.5 2.0 2.0 2.5 2.0 2.5 3.0 3.0 2.0 3.5 2.5 1.5 1.5 4.0 2.5 3.0 2.5 4.0
#>  [577] 1.5 3.0 3.0 2.0 2.5 2.0 3.5 3.0 2.5 3.5 1.5 2.0 3.0 2.5 2.5 3.5 2.0 3.0
#>  [595] 2.5 4.0 1.5 3.0 2.0 3.0 3.5 3.0 3.0 2.0 2.5 2.0 3.0 3.0 1.5 1.5 2.0 2.5
#>  [613] 1.5 3.0 2.0 2.5 2.5 1.5 2.0 2.5 2.5 2.0 2.0 4.0 3.0 2.0 4.0 1.5 1.5 2.5
#>  [631] 2.5 2.5 2.5 2.5 2.5 2.5 3.5 2.5 3.0 2.0 1.5 2.0 2.0 2.0 2.0 3.5 4.0 3.0
#>  [649] 1.5 2.0 3.5 3.0 2.5 2.5 2.0 3.0 1.0 2.0 2.0 3.0 3.5 1.5 3.5 3.0 2.5 2.5
#>  [667] 3.5 2.5 3.0 1.5 3.0 3.0 1.5 2.5 2.5 3.5 4.0 3.5 2.0 1.0 3.0 2.5 3.0 1.5
#>  [685] 2.5 3.0 3.0 2.0 1.0 3.5 2.0 3.0 1.0 3.5 2.0 3.0 2.5 2.5 2.0 3.0 2.0 2.5
#>  [703] 2.0 3.0 2.5 3.0 2.0 1.0 2.0 1.5 2.0 1.0 4.0 3.0 2.0 1.0 2.0 3.0 2.5 2.5
#>  [721] 3.5 3.0 2.5 3.0 2.0 2.0 2.5 3.0 1.5 3.5 2.5 2.5 1.5 1.5 2.0 2.5 2.0 3.0
#>  [739] 4.0 3.5 3.5 3.5 1.0 1.5 1.5 2.5 2.5 2.0 2.0 3.5 3.0 1.5 1.5 2.0 2.5 2.5
#>  [757] 2.0 3.5 2.0 2.5 2.0 2.5 3.0 2.5 2.5 3.0 3.0 3.0 3.5 1.5 3.5 3.5 1.5 3.5
#>  [775] 2.0 3.0 2.0 1.5 2.0 3.5 1.5 2.0 1.5 2.0 2.5 1.5 2.5 1.0 2.0 2.5 2.5 1.5
#>  [793] 2.0 4.0 2.5 2.5 2.5 2.5 3.0 2.5 1.0 3.0 3.5 4.0 1.5 2.5 2.0 2.0 2.0 2.0
#>  [811] 2.5 2.5 2.0 3.0 3.5 1.5 2.5 3.0 3.0 2.5 4.0 2.5 2.5 2.0 2.0 1.5 2.0 3.0
#>  [829] 2.5 3.5 4.0 1.0 3.5 2.5 2.5 3.5 2.0 2.5 3.0 4.0 3.5 2.5 1.5 2.5 3.0 3.0
#>  [847] 3.0 2.0 1.5 3.5 1.0 2.5 2.0 2.5 1.5 2.0 2.0 2.5 1.5 2.0 2.0 3.0 2.0 2.5
#>  [865] 2.0 3.5 2.5 1.5 1.5 4.0 2.0 2.5 2.5 2.5 2.5 1.5 3.0 3.0 3.0 3.5 1.5 2.0
#>  [883] 2.5 2.0 2.0 4.0 3.5 2.5 2.5 1.5 1.5 2.0 1.5 3.0 2.5 2.0 3.0 3.0 2.5 2.5
#>  [901] 2.0 1.5 1.5 3.0 3.0 3.0 1.0 1.5 2.0 3.5 1.5 3.5 3.5 1.5 3.5 3.0 2.0 2.5
#>  [919] 3.5 2.0 3.5 1.0 1.5 1.0 2.5 4.0 1.0 1.0 1.0 3.0 2.0 2.0 1.5 3.0 2.0 2.0
#>  [937] 1.0 3.0 2.5 1.0 2.5 2.5 4.0 1.5 2.5 1.0 2.0 2.5 1.0 3.0 2.5 2.5 3.0 2.5
#>  [955] 2.5 1.0 2.5 2.5 3.0 3.0 2.5 2.5 2.0 1.5 3.5 3.5 3.0 2.5 1.5 2.5 2.0 2.0
#>  [973] 2.0 2.5 2.0 2.0 2.5 2.5 2.0 2.5 3.5 3.0 3.0 3.5 2.0 3.0 3.5 2.0 1.5 3.5
#>  [991] 2.0 3.5 2.5 3.0 2.0 3.5 2.0 2.5 2.5 2.0
#>